backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
Pillow==12.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
cloudinary==1.44.2
django-cloudinary-storage==0.3.0
